feature: trainer item refactor and button
