+++ src/components/TrainersIndex/Indicators.module.css
@@ -21,5 +21,5 @@
 }
 
 .indicatorActive {
-  background-color: #87EFB6;
+  background-color: #0398fc;
 }
