+++ src/components/TrainersIndex/SearchForm.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './SearchFrom.module.css';
+import styles from './SearchForm.module.css';
 
 const SearchForm = () => (
   <form
