+++ src/components/TrainersIndex/SearchForm.module.css
@@ -3,7 +3,7 @@
   right: 40px;
   top: 40px;
   width: 300px;
-  height: 30px;
+  height: 40px;
   transition: width 0.5s ease-in;
 }
