+++ src/components/TrainersIndex/TrainerItem/Avatar.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import styles from './Avatar.module.css';
+
+const Avatar = () => (
+  <>
+    <div className={styles.avatarContainer}>
+      {/* TODO */}
+    </div>
+    <h1 className={styles.trainerName}>
+      Trainer Name
+    </h1>
+  </>
+);
+
+export default Avatar;
