+++ src/components/TrainersIndex/TrainerItem/Avatar.module.css
@@ -1,0 +1,14 @@
+.avatarContainer {
+  background-color: khaki;
+  width: 300px;
+  height: 300px;
+  margin-top: 140px;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.trainerName {
+  margin-top: 2rem;
+  margin-bottom: 2rem;
+  text-align: center;
+}
