+++ src/components/TrainersIndex/TrainerItem/Details.module.css
@@ -1,22 +1,3 @@
-.avatarContainer {
-  background-color: khaki;
-  width: 300px;
-  height: 300px;
-  margin-top: 140px;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.trainerName {
-  margin-top: 2rem;
-  margin-bottom: 2rem;
-  text-align: center;
-}
-
-.divisoryLine {
-  width: 400px;
-}
-
 .trainerDetails {
   display: flex;
   flex-wrap: wrap;
@@ -40,9 +21,4 @@
   margin-left: auto;
   margin-right: auto;
   margin-top: 50px;
-}
-
-.contactLink {
-  margin-top: 40px;
-  text-align: center;
 }
