+++ src/components/TrainersIndex/TrainerItem/index.module.css
@@ -1,0 +1,20 @@
+.divisoryLine {
+  width: 400px;
+}
+
+.contactButton {
+  margin-top: 40px;
+  text-align: center;
+  text-transform: uppercase;
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+  border: none;
+  border-radius: 8px;
+  background-color: #0398fc;
+  font-weight: bold;
+  color: white;
+  font-family: 'Roboto Slab', serif;
+  letter-spacing: 1px;
+  padding: 20px;
+}
