+++ src/index.css
@@ -16,7 +16,7 @@ ul {
 }
 
 body {
-  padding-bottom: 100px;
+  padding-bottom: 150px;
   background-color: #fcfcfc;
   font-family: 'Lora', serif;
 }
